fixing builds on mac
