+++ agent/api/task/task.go
@@ -465,30 +465,6 @@ func (task *Task) PostUnmarshalTask(cfg *config.Config,
 	return nil
 }
 
-// initializeCredentialSpecResource builds the resource dependency map for the credentialspec resource
-func (task *Task) initializeCredentialSpecResource(config *config.Config, credentialsManager credentials.Manager,
-	resourceFields *taskresource.ResourceFields) error {
-	credspecContainerMapping := task.getAllCredentialSpecRequirements()
-	credentialspecResource, err := credentialspec.NewCredentialSpecResource(task.Arn, config.AWSRegion, task.ExecutionCredentialsID,
-		credentialsManager, resourceFields.SSMClientCreator, resourceFields.S3ClientCreator, credspecContainerMapping)
-	if err != nil {
-		return err
-	}
-
-	task.AddResource(credentialspec.ResourceName, credentialspecResource)
-
-	// for every container that needs credential spec vending, it needs to wait for all credential spec resources
-	for _, container := range task.Containers {
-		if container.RequiresCredentialSpec() {
-			container.BuildResourceDependency(credentialspecResource.GetName(),
-				resourcestatus.ResourceStatus(credentialspec.CredentialSpecCreated),
-				apicontainerstatus.ContainerCreated)
-		}
-	}
-
-	return nil
-}
-
 // initNetworkMode initializes/infers the network mode for the task and assigns the result to this task's NetworkMode field.
 // ACS is streaming down this value with task payload. In case of docker bridge mode task, this value might be left empty
 // as it's the default task network mode.
@@ -2884,17 +2860,6 @@ func (task *Task) AddResource(resourceType string, resource taskresource.TaskRes
 	task.lock.Lock()
 	defer task.lock.Unlock()
 	task.ResourcesMapUnsafe[resourceType] = append(task.ResourcesMapUnsafe[resourceType], resource)
-}
-
-// requiresCredentialSpecResource returns true if at least one container in the task
-// needs a valid credentialspec resource
-func (task *Task) requiresCredentialSpecResource() bool {
-	for _, container := range task.Containers {
-		if container.RequiresCredentialSpec() {
-			return true
-		}
-	}
-	return false
 }
 
 // GetCredentialSpecResource retrieves credentialspec resource from resource map
