+++ agent/ecscni/types_windows.go
@@ -31,10 +31,6 @@ const (
 	TaskHNSNetworkNamePrefix = "task"
 	// ECSBridgeNetworkName is the name of the HNS network used as ecs-bridge.
 	ECSBridgeNetworkName = "nat"
-	// Starting with CNI plugin v0.8.0 (this PR https://github.com/containernetworking/cni/pull/698)
-	// NetworkName has to be non-empty field for network config.
-	// We do not actually make use of the field, hence passing in a placeholder string to fulfill the API spec
-	defaultNetworkName = "network-name"
 	// DefaultENIName is the name of eni interface name in the container namespace
 	DefaultENIName = "eth0"
 )
